Migrate presentation and parser modules behind stable compatibility shims.
This moves implementation ownership to presentation_pipeline/document_intelligence namespaces, rewires runtime imports to the new boundaries, and keeps legacy paper2slides/raganything paths working through shim layers.
