+++ agent/document_intelligence/__init__.py
@@ -1,0 +1,2 @@
+"""Document intelligence package boundaries."""
+
