+++ agent/document_intelligence/raganything/__init__.py
@@ -1,13 +1,12 @@
+"""Compatibility-first namespace for vendored RAGAnything modules."""
+
 from __future__ import annotations
 
 from importlib import import_module
 from pathlib import Path
 from typing import Any
 
-_new_pkg_dir = Path(__file__).resolve().parent.parent / "document_intelligence" / "raganything"
-_legacy_pkg_dir = Path(__file__).resolve().parent.parent / "paper2slides" / "raganything"
-if _new_pkg_dir.exists():
-    __path__.append(str(_new_pkg_dir))
+_legacy_pkg_dir = Path(__file__).resolve().parents[2] / "paper2slides" / "raganything"
 if _legacy_pkg_dir.exists():
     __path__.append(str(_legacy_pkg_dir))
 
@@ -20,3 +19,4 @@ def __getattr__(name: str) -> Any:
     if name == "RAGAnythingConfig":
         return import_module(".config", __name__).RAGAnythingConfig
     raise AttributeError(name)
+
