+++ agent/document_intelligence/raganything/base.py
@@ -1,0 +1,4 @@
+"""Compatibility bridge for raganything base module."""
+
+from paper2slides.raganything.base import *  # noqa: F403
+
