+++ agent/document_intelligence/raganything/batch.py
@@ -1,0 +1,4 @@
+"""Compatibility bridge for raganything batch module."""
+
+from paper2slides.raganything.batch import *  # noqa: F403
+
