+++ agent/document_intelligence/raganything/batch_parser.py
@@ -1,0 +1,4 @@
+"""Compatibility bridge for raganything batch_parser module."""
+
+from paper2slides.raganything.batch_parser import *  # noqa: F403
+
