+++ agent/document_intelligence/raganything/config.py
@@ -1,0 +1,4 @@
+"""Compatibility bridge for raganything config module."""
+
+from paper2slides.raganything.config import *  # noqa: F403
+
