+++ agent/document_intelligence/raganything/enhanced_markdown.py
@@ -1,0 +1,4 @@
+"""Compatibility bridge for raganything enhanced_markdown module."""
+
+from paper2slides.raganything.enhanced_markdown import *  # noqa: F403
+
