+++ agent/document_intelligence/raganything/modalprocessors.py
@@ -1,0 +1,4 @@
+"""Compatibility bridge for raganything modalprocessors module."""
+
+from paper2slides.raganything.modalprocessors import *  # noqa: F403
+
