+++ agent/document_intelligence/raganything/parser.py
@@ -1,0 +1,4 @@
+"""Compatibility bridge for raganything parser module."""
+
+from paper2slides.raganything.parser import *  # noqa: F403
+
