+++ agent/document_intelligence/raganything/processor.py
@@ -1,0 +1,4 @@
+"""Compatibility bridge for raganything processor module."""
+
+from paper2slides.raganything.processor import *  # noqa: F403
+
