+++ agent/document_intelligence/raganything/prompt.py
@@ -1,0 +1,4 @@
+"""Compatibility bridge for raganything prompt module."""
+
+from paper2slides.raganything.prompt import *  # noqa: F403
+
