+++ agent/document_intelligence/raganything/query.py
@@ -1,0 +1,4 @@
+"""Compatibility bridge for raganything query module."""
+
+from paper2slides.raganything.query import *  # noqa: F403
+
