+++ agent/document_intelligence/raganything/raganything.py
@@ -1,0 +1,4 @@
+"""Compatibility bridge for raganything core module."""
+
+from paper2slides.raganything.raganything import *  # noqa: F403
+
