+++ agent/document_intelligence/raganything/utils.py
@@ -1,0 +1,4 @@
+"""Compatibility bridge for raganything utils module."""
+
+from paper2slides.raganything.utils import *  # noqa: F403
+
