+++ agent/paper2slides/utils/__init__.py
@@ -1,5 +1,4 @@
-from .file_utils import save_json, load_json, save_text
-from .logging import setup_logging, log_section
+from presentation_pipeline.utils import load_json, log_section, save_json, save_text, setup_logging
 
 __all__ = [
     "save_json",
