+++ agent/paper2slides/utils/file_utils.py
@@ -1,29 +1,6 @@
-"""
-File and JSON utilities
-"""
-import json
-from pathlib import Path
-from typing import Any, Optional
+"""Legacy shim for presentation pipeline file utilities."""
 
+from presentation_pipeline.utils.file_utils import load_json, save_json, save_text
 
-def save_json(path: Path, data: Any):
-    """Save data to JSON file."""
-    path.parent.mkdir(parents=True, exist_ok=True)
-    with open(path, "w", encoding="utf-8") as f:
-        json.dump(data, f, ensure_ascii=False, indent=2, default=str)
-
-
-def load_json(path: Path) -> Optional[Any]:
-    """Load data from JSON file."""
-    if path.exists():
-        with open(path, "r", encoding="utf-8") as f:
-            return json.load(f)
-    return None
-
-
-def save_text(path: Path, text: str):
-    """Save text to file."""
-    path.parent.mkdir(parents=True, exist_ok=True)
-    with open(path, "w", encoding="utf-8") as f:
-        f.write(text)
+__all__ = ["save_json", "load_json", "save_text"]
