+++ agent/presentation_pipeline/__main__.py
@@ -1,0 +1,5 @@
+from presentation_pipeline.cli import main
+
+if __name__ == "__main__":
+    main()
+
