+++ agent/presentation_pipeline/cli.py
@@ -1,0 +1,6 @@
+"""Compatibility wrapper for the presentation pipeline CLI."""
+
+from presentation_pipeline.main import main
+
+__all__ = ["main"]
+
