+++ agent/presentation_pipeline/core/__init__.py
@@ -1,0 +1,38 @@
+"""Compatibility namespace for presentation pipeline core modules."""
+
+from pathlib import Path
+
+from paper2slides.core import (
+    STAGES,
+    detect_start_stage,
+    get_base_dir,
+    get_config_dir,
+    get_config_name,
+    get_output_dir,
+    get_plan_checkpoint,
+    get_rag_checkpoint,
+    get_summary_checkpoint,
+    get_summary_md,
+    list_outputs,
+    run_pipeline,
+)
+
+_legacy_dir = Path(__file__).resolve().parents[2] / "paper2slides" / "core"
+if _legacy_dir.exists():
+    __path__.append(str(_legacy_dir))
+
+__all__ = [
+    "get_base_dir",
+    "get_config_name",
+    "get_config_dir",
+    "get_rag_checkpoint",
+    "get_summary_checkpoint",
+    "get_summary_md",
+    "get_plan_checkpoint",
+    "get_output_dir",
+    "STAGES",
+    "detect_start_stage",
+    "run_pipeline",
+    "list_outputs",
+]
+
