+++ agent/presentation_pipeline/generator/__init__.py
@@ -1,0 +1,8 @@
+"""Compatibility namespace for presentation pipeline generator modules."""
+
+from pathlib import Path
+
+_legacy_dir = Path(__file__).resolve().parents[2] / "paper2slides" / "generator"
+if _legacy_dir.exists():
+    __path__.append(str(_legacy_dir))
+
