+++ agent/presentation_pipeline/rag/__init__.py
@@ -1,0 +1,8 @@
+"""Compatibility namespace for presentation pipeline rag modules."""
+
+from pathlib import Path
+
+_legacy_dir = Path(__file__).resolve().parents[2] / "paper2slides" / "rag"
+if _legacy_dir.exists():
+    __path__.append(str(_legacy_dir))
+
