+++ agent/presentation_pipeline/utils/slide_assets.py
@@ -1,0 +1,6 @@
+"""Compatibility bridge for slide asset extraction utilities."""
+
+from paper2slides.utils.slide_assets import SlideAssetConfig, SlideAssetExtractor
+
+__all__ = ["SlideAssetConfig", "SlideAssetExtractor"]
+
